bi-backend/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==41.0.4
Django==4.2.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-decouple==2.1
django-elasticsearch-dsl==8.0
django-environ==0.11.2
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
elastic-transport==8.13.0
elasticsearch==8.13.0
elasticsearch-dsl==8.13.0
gunicorn==21.2.0
hiredis==2.2.3
idna==3.4
kombu==5.3.6
packaging==23.2
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
setuptools==69.0.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.0.7
celery
vine==5.1.0
wcwidth==0.2.13
